Update Deploybot config.yml
I need to cut down on the Github email alerts and in my experience with js libraries when you have this many it can be much less headache to follow the practice of "if it ain't broke (or insecure) don't fix it", at least not in realtime. 

Can we try just patches weekly for a while and see if it is helpful to be less distracted by these commits?